refactor(runtime): split setup_install_state below the U16 line ceiling
Move setup-lock owner publication into setup_lock_lifecycle and managed-tree
ownership verification into setup_managed_tree_state (tree_state), keeping
setup_install_state.rs under the 800-line ceiling (507 baseline, 894 head).
