+++ vibeguard-runtime/src/setup_install_state.rs
@@ -1,9 +1,6 @@
 use crate::setup_quarantine_inventory::carry_incomplete_inventory;
-use crate::setup_support::{
-    SetupResult, home_dir, sha256_file, write_json_atomic, write_text_atomic,
-};
+use crate::setup_support::{SetupResult, home_dir, sha256_file, write_json_atomic};
 use serde_json::{Value, json};
-use std::collections::{BTreeMap, BTreeSet};
 use std::path::{Path, PathBuf};
 use std::time::{SystemTime, UNIX_EPOCH};
 
@@ -386,212 +383,6 @@ pub fn mark_complete(args: &[String]) -> SetupResult<()> {
     write_json_atomic(path, &state)
 }
 
-pub fn publish_lock_owner(args: &[String]) -> SetupResult<()> {
-    if args.len() != 3 && args.len() != 4 {
-        return Err(
-            "Usage: vibeguard-runtime setup-lock-publish-owner <lock-dir> <pid> <nonce> [reclaiming]"
-                .into(),
-        );
-    }
-    if args[1].is_empty() || args[1] == "0" || !args[1].bytes().all(|byte| byte.is_ascii_digit()) {
-        return Err("setup lock pid must be a positive decimal integer".into());
-    }
-    if args[2].is_empty() || args[2].contains(['\n', '\r']) {
-        return Err("setup lock nonce must be a non-empty single line".into());
-    }
-    let lock_dir = Path::new(&args[0]);
-    let metadata = std::fs::symlink_metadata(lock_dir)?;
-    if metadata.file_type().is_symlink() || !metadata.is_dir() {
-        return Err("setup lock path must be a regular directory".into());
-    }
-    let owner_name = args.get(3).map(String::as_str).unwrap_or("owner");
-    if owner_name != "owner" && owner_name != "reclaiming" {
-        return Err("setup lock owner name must be owner or reclaiming".into());
-    }
-    let owner = lock_dir.join(owner_name);
-    if owner.exists() || std::fs::symlink_metadata(&owner).is_ok() {
-        return Err("setup lock owner already exists".into());
-    }
-    let content = format!("pid={}\nnonce={}\n", args[1], args[2]);
-    if owner_name == "reclaiming" {
-        remove_abandoned_reclaimer_stages(lock_dir)?;
-        let staged = lock_dir.join(format!(".reclaiming.{}.{}", args[1], args[2]));
-        write_text_atomic(&staged, &content)?;
-        if let Err(link_error) = std::fs::hard_link(&staged, &owner) {
-            std::fs::remove_file(&staged).map_err(|cleanup_error| {
-                format!("{link_error}; staged reclaimer cleanup failed: {cleanup_error}")
-            })?;
-            return Err(link_error.into());
-        }
-        std::fs::remove_file(&staged)?;
-    } else {
-        write_text_atomic(&owner, &content)?;
-    }
-    #[cfg(unix)]
-    if let Err(error) = std::fs::File::open(lock_dir).and_then(|file| file.sync_all()) {
-        if let Err(cleanup_error) = std::fs::remove_file(&owner)
-            && cleanup_error.kind() != std::io::ErrorKind::NotFound
-        {
-            return Err(
-                format!("{error}; setup lock owner cleanup failed: {cleanup_error}").into(),
-            );
-        }
-        return Err(error.into());
-    }
-    Ok(())
-}
-
-fn remove_abandoned_reclaimer_stages(lock_dir: &Path) -> SetupResult<()> {
-    for entry in std::fs::read_dir(lock_dir)? {
-        let path = entry?.path();
-        let Some(name) = path.file_name().and_then(|value| value.to_str()) else {
-            return Err("setup lock contains a non-UTF-8 staged reclaimer".into());
-        };
-        if !name.starts_with(".reclaiming.") {
-            continue;
-        }
-        let metadata = std::fs::symlink_metadata(&path)?;
-        if metadata.file_type().is_symlink() || !metadata.is_file() {
-            return Err("setup lock staged reclaimer is not a regular file".into());
-        }
-        let content = std::fs::read_to_string(&path)?;
-        let mut lines = content.lines();
-        let pid = lines.next().and_then(|line| line.strip_prefix("pid="));
-        let nonce = lines.next().and_then(|line| line.strip_prefix("nonce="));
-        if lines.next().is_some()
-            || pid
-                .is_none_or(|value| value.is_empty() || !value.bytes().all(|b| b.is_ascii_digit()))
-            || nonce.is_none_or(str::is_empty)
-            || Some(name)
-                != pid
-                    .zip(nonce)
-                    .map(|(pid, nonce)| format!(".reclaiming.{pid}.{nonce}"))
-                    .as_deref()
-        {
-            return Err("setup lock staged reclaimer metadata is malformed".into());
-        }
-        std::fs::remove_file(path)?;
-    }
-    Ok(())
-}
-
-pub fn verify_managed_tree(args: &[String]) -> SetupResult<()> {
-    if args.len() != 3 && args.len() != 4 {
-        return Err(
-            "Usage: vibeguard-runtime setup-state-verify-managed-tree <state-file> <dest-dir> <source-prefix> [tracked-dest-dir]"
-                .into(),
-        );
-    }
-    let state_file = Path::new(&args[0]);
-    let dest_dir = setup_absolute_path(&expand_home(&args[1]));
-    let tracked_dest_dir = args
-        .get(3)
-        .map(|tracked| setup_absolute_path(&expand_home(tracked)))
-        .unwrap_or_else(|| dest_dir.clone());
-    println!(
-        "{}",
-        managed_tree_decision(state_file, &dest_dir, &args[2], &tracked_dest_dir)?
-    );
-    Ok(())
-}
-
-pub(crate) fn managed_tree_decision(
-    state_file: &Path,
-    dest_dir: &Path,
-    source_prefix: &str,
-    tracked_dest_dir: &Path,
-) -> SetupResult<&'static str> {
-    if !state_file.exists() {
-        return Ok("UNOWNED:no_state");
-    }
-    let state = read_state(state_file)?;
-    ensure_state_version(&state)?;
-    let source_prefix = source_prefix.trim_end_matches('/');
-    match std::fs::symlink_metadata(dest_dir) {
-        Ok(metadata) if metadata.is_dir() && !metadata.file_type().is_symlink() => {}
-        Ok(_) => return Ok("UNOWNED:path_type"),
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok("UNOWNED:missing"),
-        Err(error) => return Err(error.into()),
-    }
-
-    let files = state
-        .get("files")
-        .and_then(Value::as_object)
-        .ok_or("install-state files must be an object")?;
-    let tracked = files
-        .iter()
-        .filter_map(|(dest, info)| {
-            let expanded = setup_absolute_path(&expand_home(dest));
-            let relative = expanded.strip_prefix(tracked_dest_dir).ok()?;
-            Some((dest_dir.join(relative), info))
-        })
-        .collect::<BTreeMap<_, _>>();
-    if tracked.is_empty() {
-        return Ok("UNOWNED:not_tracked");
-    }
-
-    let mut leaves = BTreeSet::new();
-    let mut directories = BTreeSet::new();
-    collect_managed_tree_paths(dest_dir, &mut leaves, &mut directories)?;
-    for leaf in &leaves {
-        let Some(info) = tracked.get(leaf) else {
-            return Ok("UNOWNED:untracked_path");
-        };
-        if info.get("type").and_then(Value::as_str) != Some("copy") {
-            return Ok("UNOWNED:install_type");
-        }
-        let expected_source = info.get("source").and_then(Value::as_str).unwrap_or("");
-        if expected_source != source_prefix
-            && !expected_source.starts_with(&format!("{source_prefix}/"))
-        {
-            return Ok("UNOWNED:source_mismatch");
-        }
-        let Some(expected_checksum) = info.get("checksum").and_then(Value::as_str) else {
-            return Ok("UNOWNED:missing_checksum");
-        };
-        let actual_checksum = format!("sha256:{}", sha256_file(leaf)?);
-        if actual_checksum != expected_checksum {
-            return Ok("UNOWNED:checksum_mismatch");
-        }
-    }
-    if tracked.keys().any(|path| !leaves.contains(path)) {
-        return Ok("UNOWNED:missing_tracked_path");
-    }
-    if directories
-        .iter()
-        .any(|directory| !leaves.iter().any(|leaf| leaf.starts_with(directory)))
-    {
-        return Ok("UNOWNED:untracked_path");
-    }
-    Ok("OWNED")
-}
-
-fn collect_managed_tree_paths(
-    directory: &Path,
-    leaves: &mut BTreeSet<PathBuf>,
-    directories: &mut BTreeSet<PathBuf>,
-) -> SetupResult<()> {
-    for entry in std::fs::read_dir(directory)? {
-        let entry = entry?;
-        let path = entry.path();
-        let metadata = std::fs::symlink_metadata(&path)?;
-        if metadata.file_type().is_symlink() || (!metadata.is_file() && !metadata.is_dir()) {
-            return Err(format!(
-                "managed tree contains unsupported path type: {}",
-                path.display()
-            )
-            .into());
-        }
-        if metadata.is_dir() {
-            directories.insert(path.clone());
-            collect_managed_tree_paths(&path, leaves, directories)?;
-        } else {
-            leaves.insert(path);
-        }
-    }
-    Ok(())
-}
-
 pub fn list_project_hooks(args: &[String]) -> SetupResult<()> {
     if args.len() != 1 {
         return Err("Usage: vibeguard-runtime setup-state-list-project-hooks <state-file>".into());
@@ -650,7 +441,7 @@ fn read_state_or_empty(path: &Path) -> SetupResult<Value> {
     read_state(path)
 }
 
-fn ensure_state_version(state: &Value) -> SetupResult<()> {
+pub(crate) fn ensure_state_version(state: &Value) -> SetupResult<()> {
     let version = state
         .get("version")
         .and_then(Value::as_i64)
