+++ vibeguard-runtime/src/setup_lock_lifecycle.rs
@@ -150,3 +150,92 @@ fn sync_io(path: &Path) -> io::Result<()> {
     let _ = path;
     Ok(())
 }
+
+pub fn publish_lock_owner(args: &[String]) -> SetupResult<()> {
+    if args.len() != 3 && args.len() != 4 {
+        return Err(
+            "Usage: vibeguard-runtime setup-lock-publish-owner <lock-dir> <pid> <nonce> [reclaiming]"
+                .into(),
+        );
+    }
+    if args[1].is_empty() || args[1] == "0" || !args[1].bytes().all(|byte| byte.is_ascii_digit()) {
+        return Err("setup lock pid must be a positive decimal integer".into());
+    }
+    if args[2].is_empty() || args[2].contains(['\n', '\r']) {
+        return Err("setup lock nonce must be a non-empty single line".into());
+    }
+    let lock_dir = Path::new(&args[0]);
+    let metadata = std::fs::symlink_metadata(lock_dir)?;
+    if metadata.file_type().is_symlink() || !metadata.is_dir() {
+        return Err("setup lock path must be a regular directory".into());
+    }
+    let owner_name = args.get(3).map(String::as_str).unwrap_or("owner");
+    if owner_name != "owner" && owner_name != "reclaiming" {
+        return Err("setup lock owner name must be owner or reclaiming".into());
+    }
+    let owner = lock_dir.join(owner_name);
+    if owner.exists() || std::fs::symlink_metadata(&owner).is_ok() {
+        return Err("setup lock owner already exists".into());
+    }
+    let content = format!("pid={}\nnonce={}\n", args[1], args[2]);
+    if owner_name == "reclaiming" {
+        remove_abandoned_reclaimer_stages(lock_dir)?;
+        let staged = lock_dir.join(format!(".reclaiming.{}.{}", args[1], args[2]));
+        write_text_atomic(&staged, &content)?;
+        if let Err(link_error) = std::fs::hard_link(&staged, &owner) {
+            std::fs::remove_file(&staged).map_err(|cleanup_error| {
+                format!("{link_error}; staged reclaimer cleanup failed: {cleanup_error}")
+            })?;
+            return Err(link_error.into());
+        }
+        std::fs::remove_file(&staged)?;
+    } else {
+        write_text_atomic(&owner, &content)?;
+    }
+    #[cfg(unix)]
+    if let Err(error) = std::fs::File::open(lock_dir).and_then(|file| file.sync_all()) {
+        if let Err(cleanup_error) = std::fs::remove_file(&owner)
+            && cleanup_error.kind() != std::io::ErrorKind::NotFound
+        {
+            return Err(
+                format!("{error}; setup lock owner cleanup failed: {cleanup_error}").into(),
+            );
+        }
+        return Err(error.into());
+    }
+    Ok(())
+}
+
+fn remove_abandoned_reclaimer_stages(lock_dir: &Path) -> SetupResult<()> {
+    for entry in std::fs::read_dir(lock_dir)? {
+        let path = entry?.path();
+        let Some(name) = path.file_name().and_then(|value| value.to_str()) else {
+            return Err("setup lock contains a non-UTF-8 staged reclaimer".into());
+        };
+        if !name.starts_with(".reclaiming.") {
+            continue;
+        }
+        let metadata = std::fs::symlink_metadata(&path)?;
+        if metadata.file_type().is_symlink() || !metadata.is_file() {
+            return Err("setup lock staged reclaimer is not a regular file".into());
+        }
+        let content = std::fs::read_to_string(&path)?;
+        let mut lines = content.lines();
+        let pid = lines.next().and_then(|line| line.strip_prefix("pid="));
+        let nonce = lines.next().and_then(|line| line.strip_prefix("nonce="));
+        if lines.next().is_some()
+            || pid
+                .is_none_or(|value| value.is_empty() || !value.bytes().all(|b| b.is_ascii_digit()))
+            || nonce.is_none_or(str::is_empty)
+            || Some(name)
+                != pid
+                    .zip(nonce)
+                    .map(|(pid, nonce)| format!(".reclaiming.{pid}.{nonce}"))
+                    .as_deref()
+        {
+            return Err("setup lock staged reclaimer metadata is malformed".into());
+        }
+        std::fs::remove_file(path)?;
+    }
+    Ok(())
+}
